+++ claudia.json
@@ -1,0 +1,10 @@
+{
+  "lambda": {
+    "role": "verida_storage-node-executor",
+    "name": "verida_storage-node",
+    "region": "us-east-2"
+  },
+  "api": {
+    "id": "5uewrcv0fa"
+  }
+}
