+++ sample.env
@@ -1,0 +1,13 @@
+HASH_KEY=wb4f9qf6789bqjqcj0cq4897cqn890tq0
+DID_SERVER_URL="https://dids.testnet.verida.io:5001"
+DID_CACHE_DURATION=3600
+DB_PROTOCOL="http"
+DB_USER="admin"
+DB_PASS="admin"
+DB_HOST="localhost"
+DB_PORT=5984
+DB_REJECT_UNAUTHORIZED_SSL=false
+
+// alpha numeric only
+DB_PUBLIC_USER="784c2n780c9cn0789"
+DB_PUBLIC_PASS="784c2n780c9cn0789"
